+++ attribution/src/main/java/com/affise/attribution/internal/platform/InternalCrossPlatform.kt
@@ -2,19 +2,21 @@ package com.affise.attribution.internal.platform
 
 import android.net.Uri
 import com.affise.attribution.Affise
+import com.affise.attribution.AffiseApi
 
 object InternalCrossPlatform {
 
-    private fun api() = Affise.getApi()
+    private val api: AffiseApi?
+        get() = Affise.api
 
     @JvmStatic
     fun deeplink(url: String) {
-        api()?.deeplinkManager?.handleDeeplink(Uri.parse(url))
+        api?.deeplinkManager?.handleDeeplink(Uri.parse(url))
     }
 
     @JvmStatic
     fun start() {
-        api()?.sessionManager?.sessionStart()
+        api?.sessionManager?.sessionStart()
     }
 
     @JvmStatic