+++ attribution/src/main/java/com/affise/attribution/modules/AffiseHasModule.kt
@@ -1,0 +1,5 @@
+package com.affise.attribution.modules
+
+interface AffiseHasModule {
+    fun hasModule(): Boolean
+}