+++ attribution/src/main/java/com/affise/attribution/modules/AffiseModuleApiWrapper.kt
@@ -3,11 +3,20 @@ package com.affise.attribution.modules
 import com.affise.attribution.Affise
 
 abstract class AffiseModuleApiWrapper<API : AffiseModuleApi>(
-    module: AffiseModules
-) {
+    private val module: AffiseModules,
+) : AffiseHasModule {
     private var api: API? = null
 
-    protected val moduleApi: API? = api ?: Affise.Module.api<API>(module).also {
-        api = it
+    private fun <API : AffiseModuleApi> getApi(): API? {
+        return Affise.api?.moduleManager?.api(module)
+    }
+
+    protected val moduleApi: API?
+        get() = api ?: getApi<API>().also {
+            api = it
+        }
+
+    override fun hasModule(): Boolean {
+        return Affise.api?.moduleManager?.hasModule(module) ?: false
     }
 }