+++ attribution/src/main/java/com/affise/attribution/modules/appsflyer/AffiseModuleAppsFlyerApi.kt
@@ -1,0 +1,5 @@
+package com.affise.attribution.modules.appsflyer
+
+import com.affise.attribution.modules.AffiseHasModule
+
+interface AffiseModuleAppsFlyerApi : AffiseAppsFlyerApi, AffiseHasModule