+++ attribution/src/main/java/com/affise/attribution/modules/link/AffiseLink.kt
@@ -1,22 +1,15 @@
 package com.affise.attribution.modules.link
 
-import com.affise.attribution.Affise
 import com.affise.attribution.modules.AffiseModuleApiWrapper
 import com.affise.attribution.modules.AffiseModules
 
-object AffiseLink : AffiseModuleApiWrapper<AffiseLinkApi>(AffiseModules.Link) {
+internal class AffiseLink
+    : AffiseModuleApiWrapper<AffiseLinkApi>(AffiseModules.Link),
+    AffiseModuleLinkApi {
     /**
      * Module Link url Resolve
      */
-    @JvmStatic
-    fun linkResolve(url: String, callback: AffiseLinkCallback) {
-        moduleApi?.linkResolve(url, callback) ?: callback.handle("")
+    override fun resolve(url: String, callback: AffiseLinkCallback) {
+        moduleApi?.resolve(url, callback) ?: callback.handle("")
     }
-}
-
-/**
- * Module Link url Resolve
- */
-fun Affise.Module.linkResolve(url: String, callback: AffiseLinkCallback) {
-    AffiseLink.linkResolve(url, callback)
 }