+++ attribution/src/main/java/com/affise/attribution/modules/link/AffiseLinkApi.kt
@@ -3,5 +3,5 @@ package com.affise.attribution.modules.link
 import com.affise.attribution.modules.AffiseModuleApi
 
 interface AffiseLinkApi : AffiseModuleApi {
-    fun linkResolve(url: String, callback: AffiseLinkCallback)
+    fun resolve(url: String, callback: AffiseLinkCallback)
 }