+++ attribution/src/main/java/com/affise/attribution/modules/link/AffiseModuleLinkApi.kt
@@ -1,0 +1,5 @@
+package com.affise.attribution.modules.link
+
+import com.affise.attribution.modules.AffiseHasModule
+
+interface AffiseModuleLinkApi: AffiseLinkApi, AffiseHasModule