+++ attribution/src/main/java/com/affise/attribution/modules/subscription/AffiseModuleSubscriptionApi.kt
@@ -1,0 +1,5 @@
+package com.affise.attribution.modules.subscription
+
+import com.affise.attribution.modules.AffiseHasModule
+
+interface AffiseModuleSubscriptionApi: AffiseSubscriptionApi, AffiseHasModule