+++ attribution/src/main/java/com/affise/attribution/modules/subscription/AffiseSubscription.kt
@@ -6,20 +6,15 @@ import com.affise.attribution.modules.AffiseModuleApiWrapper
 import com.affise.attribution.modules.AffiseModules
 
 
-object AffiseSubscription : AffiseModuleApiWrapper<AffiseSubscriptionApi>(AffiseModules.Subscription) {
-    /**
-     * Get Subscription Module status
-     * (billingclient wrapper)
-     */
-    @JvmStatic
-    fun hasSubscriptionModule(): Boolean = moduleApi != null
+internal class AffiseSubscription
+    : AffiseModuleApiWrapper<AffiseSubscriptionApi>(AffiseModules.Subscription),
+    AffiseModuleSubscriptionApi {
 
     /**
      * Subscription Module fetch products for purchase from [productsIds]
      * (billingclient wrapper)
      */
-    @JvmStatic
-    fun fetchProducts(
+    override fun fetchProducts(
         productsIds: List<String>,
         callback: AffiseResultCallback<AffiseProductsResult>,
     ) {
@@ -32,8 +27,7 @@ object AffiseSubscription : AffiseModuleApiWrapper<AffiseSubscriptionApi>(Affise
      * Subscription Module purchase product
      * (billingclient wrapper)
      */
-    @JvmStatic
-    fun purchase(
+    override fun purchase(
         activity: Activity,
         product: AffiseProduct,
         type: AffiseProductType?,
@@ -48,8 +42,7 @@ object AffiseSubscription : AffiseModuleApiWrapper<AffiseSubscriptionApi>(Affise
      * Subscription Module purchase product
      * (billingclient wrapper)
      */
-    @JvmStatic
-    fun purchase(
+    override fun purchase(
         activity: Activity,
         productId: String,
         offerToken: String?,
@@ -63,27 +56,27 @@ object AffiseSubscription : AffiseModuleApiWrapper<AffiseSubscriptionApi>(Affise
 }
 
 @Deprecated(
-    message = "Method moved to Affise.Module",
-    replaceWith = ReplaceWith("Affise.Module.hasSubscriptionModule()")
+    message = "Method moved to Affise.Module.Subscription",
+    replaceWith = ReplaceWith("Affise.Module.Subscription.hasModule()")
 )
 fun Affise.hasSubscriptionModule(): Boolean {
-    return Affise.Module.hasSubscriptionModule()
+    return Affise.Module.Subscription.hasModule()
 }
 
 @Deprecated(
-    message = "Method moved to Affise.Module",
-    replaceWith = ReplaceWith("Affise.Module.fetchProducts(productsIds, callback)")
+    message = "Method moved to Affise.Module.Subscription",
+    replaceWith = ReplaceWith("Affise.Module.Subscription.fetchProducts(productsIds, callback)")
 )
 fun Affise.fetchProducts(
     productsIds: List<String>,
     callback: AffiseResultCallback<AffiseProductsResult>,
 ) {
-    Affise.Module.fetchProducts(productsIds, callback)
+    Affise.Module.Subscription.fetchProducts(productsIds, callback)
 }
 
 @Deprecated(
-    message = "Method moved to Affise.Module",
-    replaceWith = ReplaceWith("Affise.Module.purchase(activity, product, type, callback)")
+    message = "Method moved to Affise.Module.Subscription",
+    replaceWith = ReplaceWith("Affise.Module.Subscription.purchase(activity, product, type, callback)")
 )
 fun Affise.purchase(
     activity: Activity,
@@ -91,12 +84,12 @@ fun Affise.purchase(
     type: AffiseProductType? = null,
     callback: AffiseResultCallback<AffisePurchasedInfo>,
 ) {
-    Affise.Module.purchase(activity, product, type, callback)
+    Affise.Module.Subscription.purchase(activity, product, type, callback)
 }
 
 @Deprecated(
-    message = "Method moved to Affise.Module",
-    replaceWith = ReplaceWith("Affise.Module.purchase(activity, productId, offerToken, type, callback)")
+    message = "Method moved to Affise.Module.Subscription",
+    replaceWith = ReplaceWith("Affise.Module.Subscription.purchase(activity, productId, offerToken, type, callback)")
 )
 fun Affise.purchase(
     activity: Activity,
@@ -105,50 +98,5 @@ fun Affise.purchase(
     type: AffiseProductType? = null,
     callback: AffiseResultCallback<AffisePurchasedInfo>,
 ) {
-    Affise.Module.purchase(activity, productId, offerToken, type, callback)
-}
-
-/**
- * Get Subscription Module status
- * (billingclient wrapper)
- */
-fun Affise.Module.hasSubscriptionModule(): Boolean {
-    return AffiseSubscription.hasSubscriptionModule()
-}
-
-/**
- * Subscription Module fetch products for purchase from [productsIds]
- * (billingclient wrapper)
- */
-fun Affise.Module.fetchProducts(
-    productsIds: List<String>, callback: AffiseResultCallback<AffiseProductsResult>,
-) {
-    AffiseSubscription.fetchProducts(productsIds, callback)
-}
-
-/**
- * Subscription Module purchase product
- * (billingclient wrapper)
- */
-fun Affise.Module.purchase(
-    activity: Activity,
-    product: AffiseProduct,
-    type: AffiseProductType? = null,
-    callback: AffiseResultCallback<AffisePurchasedInfo>,
-) {
-    AffiseSubscription.purchase(activity, product, type, callback)
-}
-
-/**
- * Subscription Module purchase product
- * (billingclient wrapper)
- */
-fun Affise.Module.purchase(
-    activity: Activity,
-    productId: String,
-    offerToken: String? = null,
-    type: AffiseProductType? = null,
-    callback: AffiseResultCallback<AffisePurchasedInfo>,
-) {
-    AffiseSubscription.purchase(activity, productId, offerToken, type, callback)
+    Affise.Module.Subscription.purchase(activity, productId, offerToken, type, callback)
 }