+++ attribution/src/main/java/com/affise/attribution/modules/subscription/AffiseSubscriptionApi.kt
@@ -4,11 +4,19 @@ import android.app.Activity
 import com.affise.attribution.modules.AffiseModuleApi
 
 interface AffiseSubscriptionApi : AffiseModuleApi {
+    /**
+     * Subscription Module fetch products for purchase from [productsIds]
+     * (billingclient wrapper)
+     */
     fun fetchProducts(
         productsIds: List<String>,
         callback: AffiseResultCallback<AffiseProductsResult>,
     )
 
+    /**
+     * Subscription Module purchase product
+     * (billingclient wrapper)
+     */
     fun purchase(
         activity: Activity,
         product: AffiseProduct,
@@ -16,6 +24,10 @@ interface AffiseSubscriptionApi : AffiseModuleApi {
         callback: AffiseResultCallback<AffisePurchasedInfo>,
     )
 
+    /**
+     * Subscription Module purchase product
+     * (billingclient wrapper)
+     */
     fun purchase(
         activity: Activity,
         productId: String,